# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin swagger",
        "title": "Gin swagger",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8642",
    "basePath": "/api/v1",
    "paths": {
...
        "/task": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"Task\""
                ],
                "summary": "\"Task list\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "example:Bearer token (Bearer+space+token).",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Title (maxLength: 100)",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Specify Datetime (DateTime: 2006-01-02 15:04:05)",
                        "name": "specify_datetime",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Is Specify Time",
                        "name": "is_specify_time",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Is Complete",
                        "name": "is_complete",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Page (Please start from 1)",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit (Please start from 5 or 10)",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get task list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.PageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failed to process request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"Task\""
                ],
                "summary": "\"Create task\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "example:Bearer token (Bearer+space+token).",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title (maxLength: 100)",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Note",
                        "name": "note",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Url",
                        "name": "url",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Image",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Specify Datetime (DateTime: 2006-01-02 15:04:05)",
                        "name": "specify_datetime",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Is Specify Time",
                        "name": "is_specify_time",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Priority (Enum: [1, 2, 3])",
                        "name": "priority",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Is Complete",
                        "name": "is_complete",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failed to process request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to process request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.PageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "currentPage": {
                    "type": "integer"
                },
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "pageLimit": {
                    "type": "integer"
                },
                "pages": {
                    "description": "Total page",
                    "type": "integer"
                },
                "total": {
                    "description": "Data count",
                    "type": "integer"
                }
            }
        },
        "responses.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}